pr-watcher/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pyyaml==6.0.1
supabase==2.4.6
python-dotenv==1.0.1
